refactor(discover): Dedupe query error handling
