+++ src/plugins/data/common/query/index.ts
@@ -33,3 +33,4 @@ export * from './multi_query';
 export * from './timefilter';
 export * from './types';
 export * from './is_query';
+export * from './query_error_handler';
