+++ src/plugins/discover/public/application/components/ask_error_button/ask_error_button.tsx
@@ -6,13 +6,11 @@
 import { useCallback } from 'react';
 import { EuiIcon, EuiLink } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
+import { ASK_AI_ERROR_MESSAGE } from '../../../../../data/common';
 import { useOpenSearchDashboards } from '../../../../../opensearch_dashboards_react/public';
 import { DiscoverViewServices } from '../../../build_services';
 
 const ASK_AI_NO_RESULTS_MESSAGE = `My query on this page returned no results. Please help me adjust it and run the corrected query on the page so I can see the data. Change only one thing at a time so it is clear which change helped, and stop as soon as results appear. Check the following in order: 1) Verify the fields I referenced actually exist. 2) Check whether the filter values are valid - a value I used may simply not exist in the data; try plausible alternative values or a broader condition. 3) Only if the fields and values both look correct, the time range is the likely cause - use a backend tool to find the most recent document matching the query, then widen the time range on page gradually rather than jumping straight to a very large window.`;
-
-const ASK_AI_ERROR_MESSAGE =
-  'My query on this page failed to run with the following error: "{error}". Please review my query, fix it, and run the corrected query on the page so I can see the results.';
 
 interface AskErrorButtonProps {
   getError?: () => string | undefined;
