+++ src/plugins/discover/public/application/view_components/utils/use_search.ts
@@ -22,7 +22,7 @@ import {
 } from '../../../../../data/public';
 import { validateTimeRange } from '../../helpers/validate_time_range';
 import { updateSearchSource } from './update_search_source';
-import { extractQueryError } from './format_error';
+import { extractQueryError } from '../../../../../data/common';
 import { useIndexPattern } from './use_index_pattern';
 import { OpenSearchSearchHit } from '../../doc_views/doc_views_types';
 import { TimechartHeaderBucketInterval } from '../../components/chart/timechart_header';
